requirements.txt
beautifulsoup4
bs4
loguru
python-telegram-bot
requests
pyyaml
apscheduler
aiohttp
aiosqlite>=0.19.0
psutil
